[fix] billing ci 병목 수정
billing 통합 테스트 4개가 쓰던 Testcontainers(Rabbit/Postgres) + DynamicPropertySource 중복을 공용 베이스 테스트(AbstractBillingIntegrationTest)로 통합했습니다.
통합 테스트 패키지에서 @Testcontainers/@Container/DynamicPropertySource가 베이스 클래스에만 남도록 정리 완료.
해당 테스트 파일들 lint 에러 없음까지 확인했습니다.
