+++ services/billing-service/src/test/java/com/eevee/billingservice/integration/AbstractBillingIntegrationTest.java
@@ -1,0 +1,41 @@
+package com.eevee.billingservice.integration;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.RabbitMQContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+/**
+ * Shared Testcontainers for billing integration tests.
+ *
+ * Keeping containers in a single static holder avoids repeated startup per test class,
+ * which is a major contributor to CI runtime.
+ */
+@Testcontainers
+abstract class AbstractBillingIntegrationTest {
+
+    @Container
+    static final RabbitMQContainer rabbit = new RabbitMQContainer("rabbitmq:3.13-alpine");
+
+    @Container
+    static final PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:16-alpine")
+            .withDatabaseName("app")
+            .withUsername("app")
+            .withPassword("app");
+
+    @DynamicPropertySource
+    static void registerProps(DynamicPropertyRegistry r) {
+        r.add("spring.rabbitmq.host", rabbit::getHost);
+        r.add("spring.rabbitmq.port", rabbit::getAmqpPort);
+        r.add("spring.rabbitmq.username", () -> "guest");
+        r.add("spring.rabbitmq.password", () -> "guest");
+        r.add("spring.datasource.url", postgres::getJdbcUrl);
+        r.add("spring.datasource.username", postgres::getUsername);
+        r.add("spring.datasource.password", postgres::getPassword);
+        r.add("spring.jpa.hibernate.ddl-auto", () -> "create-drop");
+        r.add("billing.gateway.shared-secret", () -> "test-secret");
+    }
+}
+
