+++ services/billing-service/src/test/java/com/eevee/billingservice/integration/BillingRecordedEventCostOutDisabledIntegrationTest.java
@@ -9,13 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.TestPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.containers.RabbitMQContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.math.BigDecimal;
 import java.time.Instant;
@@ -30,30 +24,7 @@ import static org.awaitility.Awaitility.await;
  */
 @SpringBootTest
 @TestPropertySource(properties = "billing.rabbit.cost-out.enabled=false")
-@Testcontainers
-class BillingRecordedEventCostOutDisabledIntegrationTest {
-
-    @Container
-    static RabbitMQContainer rabbit = new RabbitMQContainer("rabbitmq:3.13-alpine");
-
-    @Container
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:16-alpine")
-            .withDatabaseName("app")
-            .withUsername("app")
-            .withPassword("app");
-
-    @DynamicPropertySource
-    static void registerProps(DynamicPropertyRegistry r) {
-        r.add("spring.rabbitmq.host", rabbit::getHost);
-        r.add("spring.rabbitmq.port", rabbit::getAmqpPort);
-        r.add("spring.rabbitmq.username", () -> "guest");
-        r.add("spring.rabbitmq.password", () -> "guest");
-        r.add("spring.datasource.url", postgres::getJdbcUrl);
-        r.add("spring.datasource.username", postgres::getUsername);
-        r.add("spring.datasource.password", postgres::getPassword);
-        r.add("spring.jpa.hibernate.ddl-auto", () -> "create-drop");
-        r.add("billing.gateway.shared-secret", () -> "test-secret");
-    }
+class BillingRecordedEventCostOutDisabledIntegrationTest extends AbstractBillingIntegrationTest {
 
     @Autowired
     private RabbitTemplate rabbitTemplate;
